Shift creation of sparse matrix out one level
The current location of the sparse matrix creation seems to create new sparse matrix for every unknown.
But the location of where the sparse matrix is created is overwritten in every iteration so only the last sparse matrix will survive.
It should therefore be equivalent to only create the sparse matrix once, outside the loop.